rename: StyledClusterMarker으로 개명
[#872]
